Serialization so far. At least login stuff is working.

+++ src/main/java/StudyBuddyApp.java
@@ -1,7 +1,9 @@
 //import UI
+import UI.LoginUI;
 import UI.UserAuthorizationUI;
 import data.persistency.UserDatabase;
 
+import javax.swing.*;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.ObjectInputStream;
@@ -10,7 +12,7 @@ import java.io.ObjectOutputStream;
 public class StudyBuddyApp {
     public static void main(String[] args){
         // deserializing the userDatabase.txt file
-        UserDatabase userDatabase = null;
+        userDatabase = new UserDatabase();
         try {
             //Creating stream to read the object
             ObjectInputStream in = new ObjectInputStream(new FileInputStream("userDatabase.txt"));
@@ -25,23 +27,47 @@ public class StudyBuddyApp {
         if (userDatabase == null) {
             userDatabase = new UserDatabase();
         }
+
+
         //initial page: user authorization
-        UserAuthorizationUI userAuthorizationPage = new UserAuthorizationUI();
+        LoginUI frame = new LoginUI();
+        frame.setTitle("Login Page");
+        frame.setVisible(true);
+        frame.setBounds(0, 0, 1440, 1000);
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        frame.setResizable(false);
 
-        userAuthorizationPage.createDialogBox("Example", 540, 300);
 
-        // serialization to userDatabase.txt file
-        try {
-            //Creating stream and writing the object
-            FileOutputStream fout = new FileOutputStream("userDatabase.txt");
-            ObjectOutputStream out = new ObjectOutputStream(fout);
-            out.writeObject(userDatabase);
-            out.flush();
-            //closing the stream
-            out.close();
-            System.out.println("successful serialization");
-        } catch (Exception e) {
-            System.out.println(e);
-        }
+
+
+        LoginUI.getFrames()[0].addWindowListener(new java.awt.event.WindowAdapter() {
+            @Override
+            public void windowClosing(java.awt.event.WindowEvent windowEvent) {
+                if (JOptionPane.showConfirmDialog(frame,
+                        "Are you sure you want to close this window?", "Close Window?",
+                        JOptionPane.YES_NO_OPTION,
+                        JOptionPane.QUESTION_MESSAGE) == JOptionPane.YES_OPTION){
+                    // serialization to userDatabase.txt file
+                    try {
+                        //Creating stream and writing the object
+                        FileOutputStream fout = new FileOutputStream("userDatabase.txt");
+                        ObjectOutputStream out = new ObjectOutputStream(fout);
+
+                        //check how many accounts are in the userDatabase
+                        int numOfAccounts = UserDatabase.getAccounts().size();
+                        System.out.println(numOfAccounts);
+
+                        out.writeObject(UserDatabase.getAccounts().size());
+                        out.flush();
+                        //closing the stream
+                        out.close();
+                        System.out.println("successful serialization");
+                    } catch (Exception e) {
+                        System.out.println(e);
+                    }
+                    System.exit(0);
+                }
+            }
+        });
     }
 }
